The findServicesWithPageSize method has been removed from the entire program.
Now, there are four methods for displaying services. One of these methods takes pageSize and elementSize as inputs and displays services based on these parameters. The findAllServices method has no conditions and shows all services. Two other methods are used for navigation.

For implementing navigation, the lazy class is used. In case of a query, this attribute requests information from the database, and with each navigation, the currentPage is updated. The default state for pageSize is set to ten.